requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# PDF Generation
reportlab==4.0.7
matplotlib==3.8.2

# Database
# SQLite is included with Python, no additional package needed

# Authentication
# Token authentication is included with DRF

# Development
python-dotenv==1.0.0
